test(lsp): sorting multiple import blocks
